QA-111: Check signed commits on community repositories
This is done through adding two flags to the chech_commits.sh script.
One for enabling the signature check, and one for enabling the Changelog check.

In the case that no flags are given, they will both be checked (The old
behaviour).

Also add the ability to pass these parameters to the
'.gitlab-ci-check-commits.yml' file through an environment variable
'MENDERTESTING_ARGS'.

Changelog: None

Signed-off-by: Ole Petter <ole.orhagen@northern.tech>